Fix bug in shardingKey overloading

+++ kraps/cluster.h
@@ -6,7 +6,7 @@
 #include "sockio.h"
 
 const size_t COORDINATOR = 0;
-const size_t ANY_NODE = (uint16_t)~0;
+const size_t ANY_NODE = (size_t)~0;
 const size_t BUF_HDR_SIZE = 16;
 const size_t MAX_PATH_LEN = 1024;
 const size_t MAX_SIZE_T = (size_t)~0;
